wip: resolve PR comments
