+++ src/components/Layout/Navbar/navbar.module.css
@@ -7,7 +7,7 @@
 
 .navbar_menu {
     color: white;
-    background: #1d1283;
+    background: var(--primary-color);
     display: flex;
     padding: 20px;
     margin: 0;
