+++ src/store/global.ts
@@ -15,12 +15,9 @@ export const global = createSlice({
     setIsLoggedIn: (state, { payload: { isLoggedIn } }) => {
       state.isLoggedIn = isLoggedIn;
     },
-    setIsLoggedOut: (state, { payload: { isLoggedIn } }) => {
-      state.isLoggedIn = isLoggedIn;
-    }
   },
 });
 
-export const { setIsLoggedIn, setIsLoggedOut } = global.actions;
+export const { setIsLoggedIn } = global.actions;
 
 export default global.reducer;
